modalHeader and emptyState folder restructure
